makes SpawnerInstance wait for proofs
A common failure in ByzCoin operation is that the instance is created on one node,
and the client asks another node for the proof, which fails, if that other node
did not get the new block yet.

This PR adds a default wait to the SpawnerInstance, and also fixes the
CalypsoWriteInstance which did not have the wait argument.