Caching counters
Instead of fetching the counters every time, keep a local cache
of the counter-states and use this cache whenever an instruction
needs to be signed.

It has some intelligence in keeping the cache up-to-date and
avoiding sending always the wrong counters. For old code, nothing
changes. For new code, it should use the
ByzCoinRPC.{updateCachedCounters,getNextCounter} methods to
sign the instructions, like shown in ClientTransaction.